feat(sdk): compose Promise host plugins
