+++ packages/sdk/README.md
@@ -11,6 +11,8 @@ const session = await opencode.sessions.create({
 })
 ```
 
+Pass imported Promise plugins in `plugins`, or register one later with `await opencode.plugin(plugin)`.
+
 The Promise API uses the same values, errors, request options, and `AsyncIterable` streams as `@opencode-ai/client`.
 
 Embedded hosts are silent by default. Set `log` to receive structured log entries:
@@ -32,6 +34,7 @@ Use the Workerd entrypoint inside a Cloudflare Durable Object. Hold one host for
 
 ```ts
 import { OpenCodeWorkerd } from "@opencode-ai/sdk/workerd"
+import myPlugin from "./my-plugin"
 
 export class OpenCodeDO {
   private readonly opencode: Promise<OpenCodeWorkerd.Interface>
@@ -40,7 +43,8 @@ export class OpenCodeDO {
     this.opencode = state.blockConcurrencyWhile(() =>
       OpenCodeWorkerd.create({
         storage: state.storage,
-        config: { content: "{}" },
+        config: { default_agent: "build" },
+        plugins: [myPlugin],
       }),
     )
   }
@@ -52,7 +56,7 @@ export class OpenCodeDO {
 }
 ```
 
-`blockConcurrencyWhile` keeps every Durable Object event out until the host is ready and resets the object if initialization fails. The retained Promise gives request handlers direct access to the same host after startup.
+`blockConcurrencyWhile` keeps every Durable Object event out until the host is ready and resets the object if initialization fails. The retained Promise gives request handlers direct access to the same host after startup. Configuration is a typed JavaScript object, and plugins are imported values bundled with the Worker.
 
 ## Effect
