+++ packages/sdk/src/effect/workerd.ts
@@ -1,20 +1,21 @@
 export * as OpenCodeWorkerd from "./workerd"
 
-import { ServerWorkerd } from "@opencode-ai/server/workerd"
 import { Layer } from "effect"
 import type { Config, Scope } from "effect"
+import { WorkerdProfile } from "../internal/workerd"
 import { OpenCode } from "./opencode"
 
-export interface CreateOptions extends Omit<ServerWorkerd.Options, "password"> {
+export type Configuration = WorkerdProfile.Configuration
+
+export interface CreateOptions extends WorkerdProfile.Options {
   readonly log?: OpenCode.CreateOptions["log"]
   readonly workspaceProviders?: OpenCode.CreateOptions["workspaceProviders"]
 }
 
-export const create = ({ log, workspaceProviders, ...options }: CreateOptions) =>
-  OpenCode.create(
-    { ...ServerWorkerd.serverOptions(options), log, workspaceProviders },
-    { overrides: ServerWorkerd.replacements(options) },
-  )
+export const create = ({ log, workspaceProviders, ...options }: CreateOptions) => {
+  const profile = WorkerdProfile.make(options)
+  return OpenCode.create({ ...profile.options, log, workspaceProviders }, { overrides: profile.replacements })
+}
 
 export const layer = (options: CreateOptions): Layer.Layer<OpenCode.Service, Config.ConfigError | Error> =>
   Layer.effect(OpenCode.Service, create(options))
